requirements.txt
bert-experimental==1.0.4
networkx==2.3
nltk==3.5
scikit-learn==0.22.2
python-magic==0.4.15
PyYAML==5.3.1
tqdm==4.45.0
py-rouge==1.1